Use a trigger-populated column for user points
This commit introduces a points column for the user model. It is
incremented and decremented whenever a verification is saved.

Furthermore, an "updatepoints" management command is added. It allows to
compute the points initially for existing data. Additionally, it is
recommended to run that management command periodically (i.e., daily) to
fix potential inconsistencies.

The points are **not** updated if the user assigned to an annotation is
changed; however, this is not an expected operation and would only
happen by manual intervention into the database.

The performance of the user explore view is much better with this
commit. The team view is significantly improved; however, a sequential
scan on the verification table is still required to calculate the points
for the last 30 days.