Add problem report page; more fixes in 500 page
The problem report page can be configured in the settings.py, using
either the PROBLEMS_URL to redirect to this page, or the PROBLEMS_TEXT
that will be displayed when a problem should be reported.

Both the base template and the 500 template user this error reporting
page for the respective links.

Further changes in the error 500 template:

- import raven.js only when raven is used
- set the image width