Fix CI code style validation
Appearently, the problem was due to an update of flake8 that included
new syntax specifications. For W503 and W504, one has to select the
preferred style now (line break before/after binary operator). Because
line breaks in this project are currently always after the operator,
W504 has been chosen to be ignored.

Additionally, a broad catch of Exception was changed to the more
specific JSONDecodeError and the unused error name was removed.