Minor behaviour change (#88)
* Fixed re-appearing scalebar

Fixed unwanted behaviour where the scalebar reappeared after being hidden by the map symbol.

* Blocked q and e in annotation mode

Buttons q and e to change the current slide are blocked in annotation mode to prevent accidentally changing the slide

* Fast annotation function

Added fast annotation function. When in annotation mode and one key is pressed, this label is directly assigned to the smallest object clicked on

* Minor bugfix for fast-annotate

Fixed bug with weired keycodes in the handleKeyPress function

* Behaviour change of fast annotate

* Behaviour change, single poly operations

Single poly operations did not work on polygons when area hit test was not activated. This is now fixed, as this was not the intended behaviour.

Co-authored-by: Christian Marzahl <christian.marzahl@gmail.com>